feat: add some feature for colour

+++ product-service/src/main/java/com/blubin/productservice/controller/BrandController.java
@@ -66,7 +66,7 @@ public class BrandController {
                     content = @Content(schema = @Schema(implementation = ErrorVm.class))),
             @ApiResponse(responseCode = "400", description = "Bad request",
                     content = @Content(schema = @Schema(implementation = ErrorVm.class)))})
-    public ResponseEntity<Void> updateBrand(@PathVariable Integer id, @Valid @RequestBody final BrandPostVm brandPostVm) {
+    public ResponseEntity<Void> updateBrand(@PathVariable Long id, @Valid @RequestBody final BrandPostVm brandPostVm) {
         brandService.update(brandPostVm, id);
         return ResponseEntity.noContent().build();
     }
@@ -78,7 +78,7 @@ public class BrandController {
                     content = @Content(schema = @Schema(implementation = ErrorVm.class))),
             @ApiResponse(responseCode = "400", description = "Bad request",
                     content = @Content(schema = @Schema(implementation = ErrorVm.class)))})
-    public ResponseEntity<Void> deleteBrand(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteBrand(@PathVariable Long id) {
         brandService.delete(id);
         return ResponseEntity.noContent().build();
     }
