+++ product-service/src/main/java/com/blubin/productservice/viewmodel/brand/BrandPostVm.java
@@ -2,7 +2,7 @@ package com.blubin.productservice.viewmodel.brand;
 
 import com.blubin.productservice.model.Brand;
 
-public record BrandPostVm(Integer id, String name, String description) {
+public record BrandPostVm(Long id, String name, String description) {
     public Brand toModel(){
         Brand brand = new Brand();
         brand.setId(id);
